Voice-Agent_communication/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlmodel==0.0.14
aiosqlite==0.19.0
httpx==0.27.0
pydantic-settings==2.5.2
python-multipart==0.0.9
websockets==12.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
